// src/app/models/user.ts
// we can use interfaces and types too for this purpose

// https://youtu.be/2Yq3za3qIp0?t=4874

export type User = {
  name: string;
  username: string;
  isEligible: boolean;
  canSwim: boolean;
  killsCount: number;
  isActive: boolean;
  isOnline: boolean;
  canDeliverSpeach: boolean;
  canFly: boolean;
  canDriveCar: boolean;
  assetAmount: number;
};